requirements.txt
PyMuPDF==1.26.7
pillow==12.0.0
pyzbar==0.1.9
requests==2.32.5
